refactor(app): keep the hop names with the provider that owns them
`FeedLatency::hop` is a borrowed string rather than a shared enum for a stated
reason: each provider's chain has different hops, and the feed-neutral module
is not supposed to know what any of them are called. The QA hook then resolved
a hop name through `LatencyHop::ALL` from inside that very module, which is the
contradiction the field's own doc comment warns against.

It moves to `feed::metatrader`, beside the provider whose chain it fakes, and
`feed::forced_latency_split` re-exports it — so the tab still reads one name and
the neutral layer is neutral again. Its tests go with it.
