+++ crates/app/src/feed/mod.rs
@@ -28,10 +28,9 @@ use quantick_engine::{Bar, Trade};
 // venue's re-export is the name the next feed author would copy.
 pub use quantick_orderbook::DepthEvent;
 
-use quantick_feed_mt5::LatencyHop;
-
 use crate::config::{FeedCapabilities, ProviderKind};
 
+pub use metatrader::forced_latency_split;
 pub use replay::{ReplayControl, ReplayLink, ReplayOptions, ReplayRequest};
 
 /// Default number of recent trades to backfill so the chart opens populated,
@@ -528,69 +527,6 @@ pub struct FeedHandle {
     pub replay: Option<ReplayLink>,
 }
 
-/// A latency split forced by `QUANTICK_FAKE_LATENCY_SPLIT`, or `None`.
-///
-/// `arrival,source,transport,hop` in milliseconds and one word, e.g.
-/// `18112,17980,132,bridge`. The readout it drives only exists while a real
-/// venue is running slowly, which is a state no screenshot can arrange and no
-/// setting reaches — so without this hook the whole cell, its hop name and its
-/// hover breakdown are invisible to anything but a human waiting for a bad day.
-///
-/// The hop is resolved through [`LatencyHop::ALL`] rather than taken as text:
-/// a name that exists is reachable here, and one that does not is refused. Two
-/// things fall out of that. The port's `hop` stays a `&'static str` with no
-/// string leaked to reach it, and a script cannot photograph a hop name the
-/// system is incapable of producing — the bargain `QUANTICK_FOOTPRINT_STYLE`
-/// already makes with `FootprintStyle::ALL`.
-///
-/// MetaTrader owns the names because it is the only provider that cuts its own
-/// chain, so the state this hook fakes is that provider's state.
-///
-/// A malformed value is logged and ignored rather than silently treated as a
-/// default — a typo in a validation script must not photograph the wrong state
-/// and call it a pass.
-#[must_use]
-pub fn forced_latency_split() -> Option<FeedLatency> {
-    let raw = std::env::var("QUANTICK_FAKE_LATENCY_SPLIT").ok()?;
-    let parsed = parse_forced_latency(raw.trim());
-    if parsed.is_none() {
-        tracing::warn!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "FAKE_LATENCY_SPLIT_REJECTED",
-            value = %raw,
-            known_hops = ?LatencyHop::ALL.map(LatencyHop::label),
-            action = "ignore",
-            "expected arrival,source,transport,hop in milliseconds"
-        );
-    }
-    parsed
-}
-
-fn parse_forced_latency(raw: &str) -> Option<FeedLatency> {
-    let mut parts = raw.split(',').map(str::trim);
-    let arrival: i64 = parts.next()?.parse().ok()?;
-    let source: i64 = parts.next()?.parse().ok()?;
-    let transport: i64 = parts.next()?.parse().ok()?;
-    let hop = match parts.next().filter(|hop| !hop.is_empty()) {
-        // Named, and the name has to be one the feed can really report.
-        Some(name) => Some(LatencyHop::from_label(name)?.label()),
-        None => None,
-    };
-    if parts.next().is_some() {
-        return None; // a fifth field means the script means something else
-    }
-    Some(FeedLatency {
-        arrival_lag_ms: arrival,
-        source_lag_ms: Some(source),
-        // One print, so the worst of the window is that print.
-        source_lag_peak_ms: Some(source),
-        transport_lag_ms: Some(transport),
-        hop,
-        prints: 1,
-    })
-}
-
 /// The initial backfill depth: `QUANTICK_BACKFILL` if it parses to a positive
 /// integer, else [`DEFAULT_BACKFILL_TARGET`].
 #[must_use]
@@ -639,48 +575,6 @@ pub fn spawn_live(
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn a_forced_split_parses_into_a_reading() {
-        // The state this hook exists for only happens while a real venue is
-        // running badly, so a validation run has to be able to ask for it.
-        let split = parse_forced_latency("18112,17980,132,MT5").expect("a well-formed hook");
-        assert_eq!(split.arrival_lag_ms, 18_112);
-        assert_eq!(split.source_lag_ms, Some(17_980));
-        assert_eq!(split.transport_lag_ms, Some(132));
-        assert_eq!(split.hop, Some("MT5"));
-        assert_eq!(split.prints, 1);
-    }
-
-    #[test]
-    fn the_hop_name_is_optional_but_the_numbers_are_not() {
-        let unnamed = parse_forced_latency("900,800,100").expect("three fields is enough");
-        assert_eq!(unnamed.hop, None);
-        assert_eq!(unnamed.arrival_lag_ms, 900);
-    }
-
-    #[test]
-    fn a_malformed_forced_split_is_rejected_not_defaulted() {
-        // A typo in a validation script must not photograph the wrong state and
-        // call it a pass, so every one of these is a refusal.
-        for bad in [
-            "",
-            "18112",
-            "18112,17980",
-            "18112,17980,x",
-            "18112,17980,132,bridge,extra",
-            "a,b,c",
-            // A hop the feed cannot report. Refused rather than shown, so a
-            // capture never claims a state the system has no way to produce.
-            "18112,17980,132,Bridge",
-            "18112,17980,132,the socket",
-        ] {
-            assert!(
-                parse_forced_latency(bad).is_none(),
-                "{bad:?} should be refused"
-            );
-        }
-    }
 
     #[test]
     fn a_provider_that_cannot_split_publishes_nothing_forever() {
